Move more elements donw to PhotosUpload
